improved error handling during attachment downloads.
