+++ apps/mail/components/mail/mail-display.tsx
@@ -37,6 +37,7 @@ import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/tooltip';
 import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover';
 import type { Sender, ParsedMessage, Attachment } from '@/types';
 import { useActiveConnection } from '@/hooks/use-connections';
+import { useAttachments } from '@/hooks/use-attachments';
 import { useBrainState } from '../../hooks/use-summary';
 import { useTRPC } from '@/providers/query-provider';
 import { useThreadLabels } from '@/hooks/use-labels';
@@ -378,9 +379,20 @@ const ActionButton = ({ onClick, icon, text, shortcut }: ActionButtonProps) => {
   );
 };
 
-const downloadAttachment = (attachment: { body: string; mimeType: string; filename: string }) => {
+const downloadAttachment = async (attachment: {
+  body: string;
+  mimeType: string;
+  filename: string;
+  attachmentId: string;
+}) => {
   try {
-    const byteCharacters = atob(attachment.body);
+    let attachmentData = attachment.body;
+
+    if (!attachmentData) {
+      throw new Error('Attachment data not found');
+    }
+
+    const byteCharacters = atob(attachmentData);
     const byteNumbers: number[] = Array(byteCharacters.length);
     for (let i = 0; i < byteCharacters.length; i++) {
       byteNumbers[i] = byteCharacters.charCodeAt(i);
@@ -398,6 +410,7 @@ const downloadAttachment = (attachment: { body: string; mimeType: string; filena
     window.URL.revokeObjectURL(url);
   } catch (error) {
     console.error('Error downloading attachment:', error);
+    toast.error('Failed to download attachment');
   }
 };
 
@@ -455,9 +468,20 @@ const handleDownloadAllAttachments =
     console.log('downloaded', subject, attachments);
   };
 
-const openAttachment = (attachment: { body: string; mimeType: string; filename: string }) => {
+const openAttachment = async (attachment: {
+  body: string;
+  mimeType: string;
+  filename: string;
+  attachmentId: string;
+}) => {
   try {
-    const byteCharacters = atob(attachment.body);
+    let attachmentData = attachment.body;
+
+    if (!attachmentData) {
+      throw new Error('Attachment data not found');
+    }
+
+    const byteCharacters = atob(attachmentData);
     const byteNumbers: number[] = Array(byteCharacters.length);
     for (let i = 0; i < byteCharacters.length; i++) {
       byteNumbers[i] = byteCharacters.charCodeAt(i);
@@ -484,6 +508,7 @@ const openAttachment = (attachment: { body: string; mimeType: string; filename: 
     }
   } catch (error) {
     console.error('Error opening attachment:', error);
+    toast.error('Failed to open attachment');
   }
 };
 
@@ -638,6 +663,7 @@ const MoreAboutQuery = ({
 const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }: Props) => {
   const [isCollapsed, setIsCollapsed] = useState<boolean>(false);
   const { data: threadData } = useThread(emailData.threadId ?? null);
+  const { data: messageAttachments } = useAttachments(emailData.id);
   //   const [unsubscribed, setUnsubscribed] = useState(false);
   //   const [isUnsubscribing, setIsUnsubscribing] = useState(false);
   const [preventCollapse, setPreventCollapse] = useState(false);
@@ -660,6 +686,7 @@ const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }:
   const { data: activeConnection } = useActiveConnection();
   const [researchSender, setResearchSender] = useState<Sender | null>(null);
   const [searchQuery, setSearchQuery] = useState<string | null>(null);
+  //   const trpc = useTRPC();
 
   const isLastEmail = useMemo(
     () => emailData.id === threadData?.latest?.id,
@@ -1077,11 +1104,11 @@ const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }:
 
             <!-- Attachments -->
             ${
-              emailData.attachments && emailData.attachments.length > 0
+              messageAttachments && messageAttachments.length > 0
                 ? `
               <div class="attachments-section">
-                <h2 class="attachments-title">Attachments (${emailData.attachments.length})</h2>
-                ${emailData.attachments
+                <h2 class="attachments-title">Attachments (${messageAttachments.length})</h2>
+                ${messageAttachments
                   .map(
                     (attachment) => `
                   <div class="attachment-item">
@@ -1491,11 +1518,11 @@ const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }:
                                 <Printer className="fill-iconLight dark:fill-iconDark mr-2 h-4 w-4" />
                                 {m['common.mailDisplay.print']()}
                               </DropdownMenuItem>
-                              {(emailData.attachments?.length ?? 0) > 0 && (
+                              {(messageAttachments?.length ?? 0) > 0 && (
                                 <DropdownMenuItem
-                                  disabled={!emailData.attachments?.length}
+                                  disabled={!messageAttachments?.length}
                                   className={
-                                    !emailData.attachments?.length
+                                    !messageAttachments?.length
                                       ? 'data-[disabled]:pointer-events-auto'
                                       : ''
                                   }
@@ -1504,7 +1531,7 @@ const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }:
                                     e.preventDefault();
                                     handleDownloadAllAttachments(
                                       emailData.subject || 'email',
-                                      emailData.attachments || [],
+                                      messageAttachments || [],
                                     )();
                                   }}
                                 >
@@ -1642,9 +1669,9 @@ const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }:
                   />
                 ) : null}
                 {/* mail attachments */}
-                {emailData?.attachments && emailData?.attachments.length > 0 ? (
+                {messageAttachments && messageAttachments.length > 0 ? (
                   <div className="mb-4 flex flex-wrap items-center gap-2 px-4 pt-4">
-                    {emailData?.attachments.map((attachment) => (
+                    {messageAttachments.map((attachment) => (
                       <div
                         key={`${attachment.filename}-${attachment.attachmentId}`}
                         className="flex"
@@ -1667,7 +1694,7 @@ const MailDisplay = ({ emailData, index, totalEmails, demo, threadAttachments }:
                         >
                           <HardDriveDownload className="text-muted-foreground dark:text-muted-foreground h-4 w-4 fill-[#FAFAFA] dark:fill-[#262626]" />
                         </button>
-                        {index < (emailData?.attachments?.length || 0) - 1 && (
+                        {index < (messageAttachments?.length || 0) - 1 && (
                           <div className="m-auto h-2 w-[1px] bg-[#E0E0E0] dark:bg-[#424242]" />
                         )}
                       </div>
