+++ apps/server/package.json
@@ -34,7 +34,7 @@
     "@hono/trpc-server": "^0.3.4",
     "@microsoft/microsoft-graph-client": "^3.0.7",
     "@microsoft/microsoft-graph-types": "^2.40.0",
-    "@modelcontextprotocol/sdk": "1.12.0",
+    "@modelcontextprotocol/sdk": "1.15.1",
     "@react-email/components": "^0.0.41",
     "@react-email/render": "1.1.0",
     "@trpc/client": "catalog:",
@@ -42,7 +42,7 @@
     "@tsndr/cloudflare-worker-jwt": "3.2.0",
     "@upstash/ratelimit": "^2.0.5",
     "@upstash/redis": "^1.34.9",
-    "agents": "0.0.93",
+    "agents": "0.0.106",
     "ai": "^4.3.13",
     "autumn-js": "catalog:",
     "base64-js": "1.5.1",
